HTTP/3
Changes:
* Client-sided resource parser now scans for related files in the format
of src="path/to/file.extension" instead of special lines containing
only the path. This allows the use of more realistic test cases
* Changed the chunk size from 100 bytes to 1000 bytes to better fit
inside a single QUIC packet

Bugfixes:
* Fixed a bug where the activeChild function in
Http3PrioritisedElementNode would use the incorrect formula for setting
the child's pseudotime. This resulted in weird issues when a node got
added later on in the connection's lifetime